etl/requirements.txt
# Core ETL libraries
pandas==2.2.0
numpy==1.26.3

# Excel/CSV handling
openpyxl==3.1.2

# Database
# sqlite3 is built-in Python

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2

# Logging
colorlog==6.8.2

# Data validation
pydantic==2.5.3
jsonschema==4.21.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
